llcppsigfetch:rm reduant test case
