+++ _xtool/internal/parser/parser_test.go
@@ -391,6 +391,76 @@ func TestNonBuiltinTypes(t *testing.T) {
 	}
 }
 
+func TestBuiltinType(t *testing.T) {
+	tests := []struct {
+		name     string
+		typ      clang.Type
+		expected ast.BuiltinType
+	}{
+		{"Void", btType(clang.TypeVoid), ast.BuiltinType{Kind: ast.Void}},
+		{"Bool", btType(clang.TypeBool), ast.BuiltinType{Kind: ast.Bool}},
+		{"Char_S", btType(clang.TypeCharS), ast.BuiltinType{Kind: ast.Char, Flags: ast.Signed}},
+		{"Char_U", btType(clang.TypeCharU), ast.BuiltinType{Kind: ast.Char, Flags: ast.Unsigned}},
+		{"Char16", btType(clang.TypeChar16), ast.BuiltinType{Kind: ast.Char16}},
+		{"Char32", btType(clang.TypeChar32), ast.BuiltinType{Kind: ast.Char32}},
+		{"WChar", btType(clang.TypeWChar), ast.BuiltinType{Kind: ast.WChar}},
+		{"Short", btType(clang.TypeShort), ast.BuiltinType{Kind: ast.Int, Flags: ast.Short}},
+		{"UShort", btType(clang.TypeUShort), ast.BuiltinType{Kind: ast.Int, Flags: ast.Short | ast.Unsigned}},
+		{"Int", btType(clang.TypeInt), ast.BuiltinType{Kind: ast.Int}},
+		{"UInt", btType(clang.TypeUInt), ast.BuiltinType{Kind: ast.Int, Flags: ast.Unsigned}},
+		{"Long", btType(clang.TypeLong), ast.BuiltinType{Kind: ast.Int, Flags: ast.Long}},
+		{"ULong", btType(clang.TypeULong), ast.BuiltinType{Kind: ast.Int, Flags: ast.Long | ast.Unsigned}},
+		{"LongLong", btType(clang.TypeLongLong), ast.BuiltinType{Kind: ast.Int, Flags: ast.LongLong}},
+		{"ULongLong", btType(clang.TypeULongLong), ast.BuiltinType{Kind: ast.Int, Flags: ast.LongLong | ast.Unsigned}},
+		{"Int128", btType(clang.TypeInt128), ast.BuiltinType{Kind: ast.Int128}},
+		{"UInt128", btType(clang.TypeUInt128), ast.BuiltinType{Kind: ast.Int128, Flags: ast.Unsigned}},
+		{"Float", btType(clang.TypeFloat), ast.BuiltinType{Kind: ast.Float}},
+		{"Half", btType(clang.TypeHalf), ast.BuiltinType{Kind: ast.Float16}},
+		{"Float16", btType(clang.TypeFloat16), ast.BuiltinType{Kind: ast.Float16}},
+		{"Double", btType(clang.TypeDouble), ast.BuiltinType{Kind: ast.Float, Flags: ast.Double}},
+		{"LongDouble", btType(clang.TypeLongDouble), ast.BuiltinType{Kind: ast.Float, Flags: ast.Long | ast.Double}},
+		{"Float128", btType(clang.TypeFloat128), ast.BuiltinType{Kind: ast.Float128}},
+		{"Complex", getComplexType(0), ast.BuiltinType{Kind: ast.Complex}},
+		{"Complex", getComplexType(ast.Double), ast.BuiltinType{Flags: ast.Double, Kind: ast.Complex}},
+		{"Complex", getComplexType(ast.Long | ast.Double), ast.BuiltinType{Flags: ast.Long | ast.Double, Kind: ast.Complex}},
+		{"Unknown", btType(clang.TypeIbm128), ast.BuiltinType{Kind: ast.Void}},
+	}
+
+	converter := &parser.Converter{}
+	converter.Convert()
+	for _, bt := range tests {
+		res := converter.ProcessBuiltinType(bt.typ)
+		if res.Kind != bt.expected.Kind {
+			t.Fatalf("%s Kind mismatch:got %d want %d, \n", bt.name, res.Kind, bt.expected.Kind)
+		}
+		if res.Flags != bt.expected.Flags {
+			t.Fatalf("%s Flags mismatch:got %d,want %d\n", bt.name, res.Flags, bt.expected.Flags)
+		}
+	}
+}
+
+// Char's Default Type in macos is signed char & in linux is unsigned char
+// So we only confirm the char's kind is char & flags is unsigned or signed
+func TestChar(t *testing.T) {
+	typ, index, transunit := GetType(&GetTypeOptions{
+		TypeCode: "char",
+		IsCpp:    false,
+	})
+	converter := &parser.Converter{}
+	expr := converter.ProcessType(typ)
+	if btType, ok := expr.(*ast.BuiltinType); ok {
+		if btType.Kind == ast.Char {
+			if btType.Flags != ast.Signed && btType.Flags != ast.Unsigned {
+				t.Fatal("Char's flags is not signed or unsigned")
+			}
+		}
+	} else {
+		t.Fatal("Char's expr is not a builtin type")
+	}
+	index.Dispose()
+	transunit.Dispose()
+}
+
 type GetTypeOptions struct {
 	TypeCode string // e.g. "char*", "char**"
 
@@ -437,4 +507,31 @@ func GetType(option *GetTypeOptions) (clang.Type, *clang.Index, *clang.Translati
 		return clang.ChildVisit_Continue
 	})
 	return typ, index, unit
+}
+
+func btType(kind clang.TypeKind) clang.Type {
+	return clang.Type{Kind: kind}
+}
+
+// get complex type from source code parsed
+func getComplexType(flag ast.TypeFlag) clang.Type {
+	var typeStr string
+	if flag&(ast.Long|ast.Double) == (ast.Long | ast.Double) {
+		typeStr = "long double"
+	} else if flag&ast.Double != 0 {
+		typeStr = "double"
+	} else {
+		typeStr = "float"
+	}
+
+	code := fmt.Sprintf("#include <complex.h>\n%s complex", typeStr)
+
+	// todo(zzy):free index and unit after test
+	typ, _, _ := GetType(&GetTypeOptions{
+		TypeCode:       code,
+		ExpectTypeKind: clang.TypeComplex,
+		IsCpp:          false,
+	})
+
+	return typ
 }
