--- _xtool/llcppsigfetch/parse/cvt_test/complex_test/forwarddecl_test/forwarddecl.go
@@ -1,30 +1,0 @@
-package main
-
-import (
-	"github.com/goplus/llcppg/_xtool/llcppsigfetch/parse"
-	test "github.com/goplus/llcppg/_xtool/llcppsigfetch/parse/cvt_test"
-	llcppg "github.com/goplus/llcppg/config"
-)
-
-func main() {
-	TestForwardDecl()
-	TestForwardDeclCrossFile()
-}
-
-func TestForwardDecl() {
-	test.RunTestWithConfig(&parse.Config{
-		Conf: &llcppg.Config{
-			Include: []string{"forwarddecl.h"},
-			CFlags:  "-I./hfile/",
-		},
-	})
-}
-
-func TestForwardDeclCrossFile() {
-	test.RunTestWithConfig(&parse.Config{
-		Conf: &llcppg.Config{
-			Include: []string{"def.h"},
-			CFlags:  "-I./hfile/",
-		},
-	})
-}
