--- _xtool/llcppsigfetch/parse/cvt_test/complex_test/forwarddecl_test/hfile/def.h
@@ -1,3 +1,0 @@
-#include "impl.h"
-typedef struct foo foo;
-void f(foo *f);
