--- _xtool/llcppsigfetch/parse/cvt_test/complex_test/forwarddecl_test/hfile/impl.h
@@ -1,4 +1,0 @@
-struct foo
-{
-    int a;
-};
