--- _xtool/llcppsigfetch/parse/cvt_test/complex_test/preprocess_test/hfile/compat.h
@@ -1,4 +1,0 @@
-#ifndef COMPAT_H
-#define COMPAT_H
-typedef A B;
-#endif
