--- _xtool/llcppsigfetch/parse/cvt_test/complex_test/preprocess_test/hfile/main.h
@@ -1,8 +1,0 @@
-typedef struct A
-{
-    int a;
-    int b;
-    int c;
-} A;
-#include "compat.h"
-typedef B C;
