--- _xtool/llcppsigfetch/parse/cvt_test/decl_test/enum_test/llgo.expect
@@ -1,258 +1,0 @@
-#stdout
-TestEnumDecl Case 1:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"EnumTypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	null,
-				"Type":	{
-					"_Type":	"EnumType",
-					"Items":	[{
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"a"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"0"
-							}
-						}, {
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"b"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"1"
-							}
-						}, {
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"c"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"2"
-							}
-						}]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestEnumDecl Case 2:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"EnumTypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"Foo"
-				},
-				"Type":	{
-					"_Type":	"EnumType",
-					"Items":	[{
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"a"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"0"
-							}
-						}, {
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"b"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"1"
-							}
-						}, {
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"c"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"2"
-							}
-						}]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestEnumDecl Case 3:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"EnumTypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"Foo"
-				},
-				"Type":	{
-					"_Type":	"EnumType",
-					"Items":	[{
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"a"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"1"
-							}
-						}, {
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"b"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"2"
-							}
-						}, {
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"c"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"4"
-							}
-						}]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestEnumDecl Case 4:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"EnumTypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"Foo"
-				},
-				"Type":	{
-					"_Type":	"EnumType",
-					"Items":	[{
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"a"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"1"
-							}
-						}, {
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"b"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"2"
-							}
-						}, {
-							"_Type":	"EnumItem",
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"c"
-							},
-							"Value":	{
-								"_Type":	"BasicLit",
-								"Kind":	0,
-								"Value":	"3"
-							}
-						}]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-
-#stderr
-
-#exit 0
