--- _xtool/llcppsigfetch/parse/cvt_test/decl_test/scope_test/llgo.expect
@@ -1,316 +1,0 @@
-#stdout
-TestScope Case 1:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"FuncDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"foo"
-				},
-				"MangledName":	"_Z3foov",
-				"Type":	{
-					"_Type":	"FuncType",
-					"Params":	{
-						"_Type":	"FieldList",
-						"List":	null
-					},
-					"Ret":	{
-						"_Type":	"BuiltinType",
-						"Kind":	0,
-						"Flags":	0
-					}
-				},
-				"IsInline":	false,
-				"IsStatic":	false,
-				"IsConst":	false,
-				"IsExplicit":	false,
-				"IsConstructor":	false,
-				"IsDestructor":	false,
-				"IsVirtual":	false,
-				"IsOverride":	false
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestScope Case 2:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"FuncDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	{
-					"_Type":	"Ident",
-					"Name":	"a"
-				},
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"foo"
-				},
-				"MangledName":	"_ZN1a3fooEv",
-				"Type":	{
-					"_Type":	"FuncType",
-					"Params":	{
-						"_Type":	"FieldList",
-						"List":	null
-					},
-					"Ret":	{
-						"_Type":	"BuiltinType",
-						"Kind":	0,
-						"Flags":	0
-					}
-				},
-				"IsInline":	false,
-				"IsStatic":	false,
-				"IsConst":	false,
-				"IsExplicit":	false,
-				"IsConstructor":	false,
-				"IsDestructor":	false,
-				"IsVirtual":	false,
-				"IsOverride":	false
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestScope Case 3:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"FuncDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	{
-					"_Type":	"ScopingExpr",
-					"X":	{
-						"_Type":	"Ident",
-						"Name":	"b"
-					},
-					"Parent":	{
-						"_Type":	"Ident",
-						"Name":	"a"
-					}
-				},
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"foo"
-				},
-				"MangledName":	"_ZN1a1b3fooEv",
-				"Type":	{
-					"_Type":	"FuncType",
-					"Params":	{
-						"_Type":	"FieldList",
-						"List":	null
-					},
-					"Ret":	{
-						"_Type":	"BuiltinType",
-						"Kind":	0,
-						"Flags":	0
-					}
-				},
-				"IsInline":	false,
-				"IsStatic":	false,
-				"IsConst":	false,
-				"IsExplicit":	false,
-				"IsConstructor":	false,
-				"IsDestructor":	false,
-				"IsVirtual":	false,
-				"IsOverride":	false
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestScope Case 4:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"TypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"a"
-				},
-				"Type":	{
-					"_Type":	"RecordType",
-					"Tag":	3,
-					"Fields":	{
-						"_Type":	"FieldList",
-						"List":	null
-					},
-					"Methods":	[{
-							"_Type":	"FuncDecl",
-							"Loc":	{
-								"_Type":	"Location",
-								"File":	"./temp.h"
-							},
-							"Doc":	null,
-							"Parent":	{
-								"_Type":	"Ident",
-								"Name":	"a"
-							},
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"foo"
-							},
-							"MangledName":	"_ZN1a3fooEv",
-							"Type":	{
-								"_Type":	"FuncType",
-								"Params":	{
-									"_Type":	"FieldList",
-									"List":	null
-								},
-								"Ret":	{
-									"_Type":	"BuiltinType",
-									"Kind":	0,
-									"Flags":	0
-								}
-							},
-							"IsInline":	false,
-							"IsStatic":	false,
-							"IsConst":	false,
-							"IsExplicit":	false,
-							"IsConstructor":	false,
-							"IsDestructor":	false,
-							"IsVirtual":	false,
-							"IsOverride":	false
-						}]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestScope Case 5:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"TypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	{
-					"_Type":	"Ident",
-					"Name":	"a"
-				},
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"b"
-				},
-				"Type":	{
-					"_Type":	"RecordType",
-					"Tag":	3,
-					"Fields":	{
-						"_Type":	"FieldList",
-						"List":	null
-					},
-					"Methods":	[{
-							"_Type":	"FuncDecl",
-							"Loc":	{
-								"_Type":	"Location",
-								"File":	"./temp.h"
-							},
-							"Doc":	null,
-							"Parent":	{
-								"_Type":	"ScopingExpr",
-								"X":	{
-									"_Type":	"Ident",
-									"Name":	"b"
-								},
-								"Parent":	{
-									"_Type":	"Ident",
-									"Name":	"a"
-								}
-							},
-							"Name":	{
-								"_Type":	"Ident",
-								"Name":	"foo"
-							},
-							"MangledName":	"_ZN1a1b3fooEv",
-							"Type":	{
-								"_Type":	"FuncType",
-								"Params":	{
-									"_Type":	"FieldList",
-									"List":	null
-								},
-								"Ret":	{
-									"_Type":	"BuiltinType",
-									"Kind":	0,
-									"Flags":	0
-								}
-							},
-							"IsInline":	false,
-							"IsStatic":	false,
-							"IsConst":	false,
-							"IsExplicit":	false,
-							"IsConstructor":	false,
-							"IsDestructor":	false,
-							"IsVirtual":	false,
-							"IsOverride":	false
-						}]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-
-#stderr
-
-#exit 0
