--- _xtool/llcppsigfetch/parse/cvt_test/decl_test/struct_test/llgo.expect
@@ -1,400 +1,0 @@
-#stdout
-TestStructDecl Case 1:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"TypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	null,
-				"Type":	{
-					"_Type":	"RecordType",
-					"Tag":	0,
-					"Fields":	{
-						"_Type":	"FieldList",
-						"List":	[{
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"BuiltinType",
-									"Kind":	6,
-									"Flags":	0
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"a"
-									}]
-							}]
-					},
-					"Methods":	[]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestStructDecl Case 2:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"TypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"A"
-				},
-				"Type":	{
-					"_Type":	"RecordType",
-					"Tag":	0,
-					"Fields":	{
-						"_Type":	"FieldList",
-						"List":	[{
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"BuiltinType",
-									"Kind":	6,
-									"Flags":	0
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"a"
-									}]
-							}, {
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"BuiltinType",
-									"Kind":	6,
-									"Flags":	0
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"b"
-									}]
-							}]
-					},
-					"Methods":	[]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestStructDecl Case 3:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"TypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"A"
-				},
-				"Type":	{
-					"_Type":	"RecordType",
-					"Tag":	0,
-					"Fields":	{
-						"_Type":	"FieldList",
-						"List":	[{
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"BuiltinType",
-									"Kind":	6,
-									"Flags":	0
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"a"
-									}]
-							}, {
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"BuiltinType",
-									"Kind":	6,
-									"Flags":	0
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"b"
-									}]
-							}]
-					},
-					"Methods":	[]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestStructDecl Case 4:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"TypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"A"
-				},
-				"Type":	{
-					"_Type":	"RecordType",
-					"Tag":	0,
-					"Fields":	{
-						"_Type":	"FieldList",
-						"List":	[{
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"BuiltinType",
-									"Kind":	6,
-									"Flags":	0
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"a"
-									}]
-							}, {
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"PointerType",
-									"X":	{
-										"_Type":	"FuncType",
-										"Params":	{
-											"_Type":	"FieldList",
-											"List":	[{
-													"_Type":	"Field",
-													"Type":	{
-														"_Type":	"BuiltinType",
-														"Kind":	6,
-														"Flags":	0
-													},
-													"Doc":	null,
-													"Comment":	null,
-													"IsStatic":	false,
-													"Access":	0,
-													"Names":	null
-												}, {
-													"_Type":	"Field",
-													"Type":	{
-														"_Type":	"BuiltinType",
-														"Kind":	6,
-														"Flags":	0
-													},
-													"Doc":	null,
-													"Comment":	null,
-													"IsStatic":	false,
-													"Access":	0,
-													"Names":	null
-												}]
-										},
-										"Ret":	{
-											"_Type":	"BuiltinType",
-											"Kind":	6,
-											"Flags":	0
-										}
-									}
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"Foo"
-									}]
-							}]
-					},
-					"Methods":	[]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-TestStructDecl Case 5:
-{
-	"File":	{
-		"_Type":	"File",
-		"decls":	[{
-				"_Type":	"TypeDecl",
-				"Loc":	{
-					"_Type":	"Location",
-					"File":	"./temp.h"
-				},
-				"Doc":	null,
-				"Parent":	null,
-				"Name":	{
-					"_Type":	"Ident",
-					"Name":	"Person"
-				},
-				"Type":	{
-					"_Type":	"RecordType",
-					"Tag":	0,
-					"Fields":	{
-						"_Type":	"FieldList",
-						"List":	[{
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"BuiltinType",
-									"Kind":	6,
-									"Flags":	0
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"age"
-									}]
-							}, {
-								"_Type":	"Field",
-								"Type":	{
-									"_Type":	"RecordType",
-									"Tag":	0,
-									"Fields":	{
-										"_Type":	"FieldList",
-										"List":	[{
-												"_Type":	"Field",
-												"Type":	{
-													"_Type":	"BuiltinType",
-													"Kind":	6,
-													"Flags":	0
-												},
-												"Doc":	null,
-												"Comment":	null,
-												"IsStatic":	false,
-												"Access":	1,
-												"Names":	[{
-														"_Type":	"Ident",
-														"Name":	"year"
-													}]
-											}, {
-												"_Type":	"Field",
-												"Type":	{
-													"_Type":	"BuiltinType",
-													"Kind":	6,
-													"Flags":	0
-												},
-												"Doc":	null,
-												"Comment":	null,
-												"IsStatic":	false,
-												"Access":	1,
-												"Names":	[{
-														"_Type":	"Ident",
-														"Name":	"day"
-													}]
-											}, {
-												"_Type":	"Field",
-												"Type":	{
-													"_Type":	"BuiltinType",
-													"Kind":	6,
-													"Flags":	0
-												},
-												"Doc":	null,
-												"Comment":	null,
-												"IsStatic":	false,
-												"Access":	1,
-												"Names":	[{
-														"_Type":	"Ident",
-														"Name":	"month"
-													}]
-											}]
-									},
-									"Methods":	[]
-								},
-								"Doc":	null,
-								"Comment":	null,
-								"IsStatic":	false,
-								"Access":	1,
-								"Names":	[{
-										"_Type":	"Ident",
-										"Name":	"birthday"
-									}]
-							}]
-					},
-					"Methods":	[]
-				}
-			}],
-		"includes":	[],
-		"macros":	[]
-	},
-	"FileMap":	{
-		"temp.h":	{
-			"FileType":	1
-		}
-	}
-}
-
-
-#stderr
-
-#exit 0
