--- _xtool/llcppsigfetch/parse/cvt_test/decl_test/union_test/union.go
@@ -1,29 +1,0 @@
-package main
-
-import test "github.com/goplus/llcppg/_xtool/llcppsigfetch/parse/cvt_test"
-
-func main() {
-	TestUnionDecl()
-}
-
-func TestUnionDecl() {
-	testCases := []string{
-		`union {
-			int a;
-			int b;
-		};`,
-		`union A {
-			int a;
-			int b;
-		};`,
-		`union OuterUnion {
-    		int i;
-    		float f;
-    		union {
-        		int c;
-        		short s;
-    		} inner;
-		};`,
-	}
-	test.RunTest("TestUnionDecl", testCases)
-}
